refactor: Add ReporterCore target to make reuse easier (#21)
This change introduces a new 'ReporterCore' target which is used by the
'reporter' command and can eaaily be imported in tests. Right now the
new target is nearly empty (it only includes `Shell`) but most
functionality can be moved here. This is preparatory work for fleshing
out the tests and adding an Xcode-based signed macOS build.