ci: Update Swift version (#63)
This change switches to using `swift-actions/setup-swift@v2` for
managing the Swift toolchain and tracks the latest Swift version. Using
mise seems to be an unreliable mechanism for managing Swift and has
broken the build due to Swift crashes, motivating this change.
Ultimately it seems we should be using Swiftly for this.